allow programmitic usage of update functionality
by separating the presentation layer concerns, or the CLI concerns,
from the application layer concerns, I was able to relatively easily
expose another presentation layer, the index.js file. The index.js
file can be used by implementing applications to programmtically
run the functionality that was previously only exposed via the CLI.

In this commit, the only logic that changed was on line 9 of the new
app/update.js file, where a log statement was changed to throwing an
error. Otherwise, no changes were made. The existing code was just
moved around.